tests/wpt/web-platform-tests/css/selectors/parsing/parse-focus-visible.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Selectors: :focus-visible pseudo-class</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="https://drafts.csswg.org/selectors-4/#the-focus-visible-pseudo">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test_valid_selector(':focus-visible');
  test_valid_selector('a:focus-visible');
  test_valid_selector(':focus:not(:focus-visible)');
</script>
